tolerance datatype changed (#641)

+++ contracts/Core/parameters/Governance.sol
@@ -140,7 +140,7 @@ contract Governance is Initializable, ACL, Constants {
         stakeManagerParams.setEpochLimitForUpdateCommission(_epochLimitForUpdateCommission);
     }
 
-    function setMaxTolerance(uint8 _maxTolerance) external onlyRole(GOVERNER_ROLE) {
+    function setMaxTolerance(uint16 _maxTolerance) external onlyRole(GOVERNER_ROLE) {
         require(_maxTolerance <= BASE_DENOMINATOR, "Slash nums addtion exceeds 10000");
         emit ParameterChanged(msg.sender, "maxTolerance", _maxTolerance, block.timestamp);
         assetManagerParams.setMaxTolerance(_maxTolerance);
